refactor episode a little bit
