+++ src/pages/index.js
@@ -4,7 +4,7 @@ import IndexCSSModule from './index.module.css'
 import { graphql } from 'gatsby'
 import { I18n } from 'react-i18next'
 import { Link, withI18next } from 'gatsby-plugin-i18next'
-import Episode from '../components/episode'
+import Episode from '../components/episode/episode'
 import RandomEpisodes from '../components/randomEpisodes'
 import SubscribeBtn from '../components/subscribeBtn'
 import Support from '../components/support'
